🚚 Rename: MovieDetail과 그 스타일 파일을 MovieDetail 폴더로 이동
