+++ src/components/MovieDetail/MovieDetail.styles.jsx
